unnecessary_fold to its own module
